+++ app/join-project/[projectId]/[hash]/page.tsx
@@ -1,80 +1,19 @@
-'use client';
+import { Metadata } from 'next';
+import JoinProjectClient from './join-project-client';
 
-import { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { validateProjectInviteHash } from '@/lib/utils/invite-links';
-import { supabase } from '@/lib/supabase';
-import { LoadingScreen } from '@/components/ui/loading-screen';
-import { ErrorScreen } from '@/components/ui/error-screen';
+export const metadata: Metadata = {
+  title: 'Join Project',
+  description: 'Join a project team',
+};
 
-export default function JoinProject({
-  params: { projectId, hash },
-}: {
-  params: { projectId: string; hash: string };
-}) {
-  const router = useRouter();
-  const [error, setError] = useState<string>();
-  const [loading, setLoading] = useState(true);
+interface Props {
+  params: Promise<{
+    projectId: string;
+    hash: string;
+  }>;
+}
 
-  useEffect(() => {
-    const joinProject = async () => {
-      try {
-        if (!validateProjectInviteHash(projectId, hash)) {
-          setError('Invalid invite link');
-          setLoading(false);
-          return;
-        }
-
-        const { data: { user } } = await supabase.auth.getUser();
-        if (!user) {
-          router.push(`/login?returnUrl=/join-project/${projectId}/${hash}`);
-          return;
-        }
-
-        const { data: project, error: projectError } = await supabase
-          .from('projects')
-          .select('teammates')
-          .eq('id', projectId)
-          .single();
-
-        if (projectError || !project) {
-          setError('Project not found');
-          setLoading(false);
-          return;
-        }
-
-        const teammates = project.teammates || [];
-        if (teammates.includes(user.email)) {
-          router.push(`/events/${projectId}/projects/public/${projectId}`);
-          return;
-        }
-
-        const { error: updateError } = await supabase
-          .from('projects')
-          .update({
-            teammates: [...teammates, user.email],
-          })
-          .eq('id', projectId);
-
-        if (updateError) {
-          setError('Failed to join project');
-          setLoading(false);
-          return;
-        }
-
-        router.push(`/events/${projectId}/projects/public/${projectId}`);
-      } catch (err) {
-        setError('An unexpected error occurred');
-        setLoading(false);
-      }
-    };
-
-    joinProject();
-  }, [projectId, hash, router]);
-
-  if (error) {
-    return <ErrorScreen message={error} />;
-  }
-
-  return <LoadingScreen />;
+export default async function JoinProjectPage({ params }: Props) {
+  const resolvedParams = await params;
+  return <JoinProjectClient projectId={resolvedParams.projectId} hash={resolvedParams.hash} />;
 }
