+++ package.json
@@ -50,16 +50,16 @@
     "html5-qrcode": "^2.3.8",
     "latest": "^0.2.0",
     "lucide-react": "^0.462.0",
-    "next": "^15.2.4",
+    "next": "14.1.0",
     "next-pwa": "^5.6.0",
     "next-themes": "^0.4.4",
     "nodemailer": "^6.9.9",
     "openai": "^4.85.0",
     "papaparse": "^5.5.2",
     "qrcode.react": "^4.1.0",
-    "react": "^18.3.1",
+    "react": "18.2.0",
     "react-date-range": "^2.0.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.53.2",
     "react-hot-toast": "^2.5.2",
     "react-markdown": "^10.1.0",
@@ -82,7 +82,7 @@
     "@types/react-dom": "^18",
     "env-cmd": "^10.1.0",
     "eslint": "^8",
-    "eslint-config-next": "15.0.3",
+    "eslint-config-next": "14.1.0",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
